refactor: auth changes
